Fixed IntMath.max(int...) wrongly returning float.

+++ src/de/diddiz/utils/IntMath.java
@@ -54,7 +54,7 @@ public final class IntMath
 	 * 
 	 * @throws ArrayIndexOutOfBoundsException When array is empty.
 	 */
-	public static float max(int... arr) {
+	public static int max(int... arr) {
 		int max = arr[0];
 		for (int i = 1; i < arr.length; i++)
 			if (arr[i] > max)
